% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertformats.R
\name{ah2sp}
\alias{ah2sp}
\title{Convert an alphahull to a SpatialPolygonsDataFrame}
\usage{
ah2sp(x, increment = 360, rnd = 10, proj4string = CRS(as.character(NA)))
}
\arguments{
\item{x}{An object of class ahull from the alphahull package}

\item{increment}{A single numeric value for the number of increments with which to describe the circumference of a circle.}

\item{rnd}{A single numeric value controlling the precision of the estimate of the arcs (the default is usually sensible).}

\item{proj4string}{Projection parameters if the alphahull is based on geographic locations.}
}
\value{
An object of class SpatialPolygonsDataFrame
}
\description{
Convert an ahull object from the alphahull package to a SpatialPolygonsDataFramefor wider use.
}
\examples{
##Example point set and its alphahull
n <- 300
theta <- runif(n, 0,2 * pi)
r <- sqrt(runif(n, 0.25^2, 0.5^2))
x <- cbind(0.5 + r*cos(theta), 0.5 + r*sin(theta))
ahullobj <- ahull(x, alpha=0.1)
ahobj <- ah2sp(ahullobj)
## Plot
par(mfrow=c(1,2), pty="s")
plot(ahullobj, main="ahull object")
plot(ahobj, col="lightgrey", pbg="white", axes=TRUE, main="sp object")
points(x)
}
