% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{arrangeSymbols}
\alias{arrangeSymbols}
\title{Rearrange the location of those symbols that overalap with each other on a map.}
\usage{
arrangeSymbols(x, plotsizes, method = "buffers1")
}
\arguments{
\item{x}{a SpatialPoints* object}

\item{plotsizes}{a vector of sizes (typically radii or perhaps half a bounding box) for the plotted symbols}
}
\value{
matrix of new suggested x and y locations for the centre of each plot symbol.
}
\description{
This function takes the orginal plot locations and suggests new locations that avoid each other, given a stated plotting size for each symbol.
}
