% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{flylines}
\alias{flylines}
\title{Produce flyout lines for where mapped pies and other charts need to avoid overlapping.}
\usage{
flylines(x, y, xplot, yplot)
}
\arguments{
\item{x}{a vector of x coordinates for the original locations}

\item{y}{a vector of y coordinates for the original locations}

\item{xplot}{a vector of x coordinates for where the charts will actually be plotted.}

\item{yplot}{a vector of y coordinates for where the charts will actually be plotted.}
}
\value{
SpatialLines
}
\description{
This function takes the orginal plot locations and the suggested new plot locations created by arrangeSymbols and creates lines between them that can be plotted.
}
