% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonprocessing.R
\name{gEdgePoints}
\alias{gEdgePoints}
\title{Adjust a polygonal study area to incorporate points near edges}
\usage{
gEdgePoints(x, pts, width)
}
\arguments{
\item{x}{input study area of class SpatialPolygons*}

\item{pts}{observations of class SpatialPoints*}

\item{width}{width to buffer the points}
}
\value{
An object of class SpatialPolygonsDataFrame
}
\description{
Function to buffer an existing polygonal study area to create a revised study area that includes all known points.
}
\examples{
utm34n <- CRS("+init=epsg:32634") # Greek UTM (west)
## Create some points within and beyond polygon b2
b1 <- basicbox(c(681900,4010900), c(682500,4011400), proj4string=utm34n)
b2 <- basicbox(c(681920,4010920), c(682480,4011380), proj4string=utm34n)
set.seed(123)
pts <- spsample(b1,100, type="random")
## New polygon to include all points
b2b <- gEdgePoints(b2,pts,25)
plot(b2b, border="red")
points(pts, pch=19, cex=0.3, col="blue")
plot(b2, add=TRUE, border="black", lty="dotted")
}
