% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{mapElements}
\alias{mapElements}
\title{Addition of both a north arrow and scalebar to map}
\usage{
mapElements(x, y, scalesize, scalelabel, cex.label = 0.75,
  type.north = "simple", lwd.north = 0.5, offset.north = height.scale
  * 1.5, r.north = scalesize/5, bkgrd.north = "white",
  fill.north = "black", offset.scale = height.scale,
  height.scale = scalesize/8, border.scale = "black",
  col.scale = "white", lwd.scale = 0.5)
}
\arguments{
\item{x}{x coordinate for the centre of the added elements (the centre of the text label)}

\item{y}{y coordinate for the centre of the added elements (the centre of the text label)}

\item{scalesize}{width of the scalebar in map units (this function assumes projcted units but will not throw a warning if unprojected units are used.}

\item{scalelabel}{text label for the state width of the scalebar (it requires the user to enter this and is not automatically generated from scalesize)}
}
\value{
a plotted north arrow and scale-bar
}
\description{
This is a helper function to add both north arrow and scalebar to a map.
}
