% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertformats.R
\name{ngrResolution}
\alias{ngrResolution}
\title{Change the resolution of a British National Grid Reference.}
\usage{
ngrResolution(x, figures)
}
\arguments{
\item{x}{A character string with a British National Grid Reference}

\item{figures}{The desired precision (an even whole number)}
}
\value{
A character string with a new British National Grid Reference
}
\description{
Increase of decrease the resolution of a British National Grid Reference.
}
\examples{
x <- "SJ456782"
ngrResolution(x,4)
ngrResolution(x,8)
ngrResolution(x,7)
}
