% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertformats.R
\name{ngrZeroStrip}
\alias{ngrZeroStrip}
\title{Remove trailing zeroes in National Grid Reference}
\usage{
ngrZeroStrip(x)
}
\arguments{
\item{x}{A string containing a valid National Grid Reference}
}
\value{
A string containing a valid National Grid Reference
}
\description{
Remove trailing zeroes in National Grid Reference when present in both x and y
}
\examples{
ngrZeroStrip("ST46409000")
ngrZeroStrip("ST46409001")
}
