% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{northArrow}
\alias{northArrow}
\title{Produce a simple north arrow}
\usage{
northArrow(x, y, r, type = "simple", fill = "black", bkgrd = NA, ...)
}
\arguments{
\item{x}{x coordinate for the centre of the bounding circle of the north arrow}

\item{y}{y coordinate for the centre of the bounding circle of the north arrow}

\item{r}{Radius of the bounding circle of the north arrow}
}
\value{
a plotted polygon
}
\description{
This function draw a simple north arrow of a certain size and at a certain map location.
}
