% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{ribbonplot}
\alias{ribbonplot}
\title{Handle the positioning and labelling of a (spatstat) colour ramp more easily}
\usage{
ribbonplot(
  x,
  y,
  xprop,
  yprop,
  zlim,
  cm = NULL,
  steps = 4,
  sigdigits = NULL,
  rounding = NULL,
  vertical = TRUE,
  ...
)
}
\arguments{
\item{x}{y-axis location of the colourramp lower-left corner.}

\item{xprop}{proportion of plot width to be allocated to the colour ramp width}

\item{yprop}{proportion of plot height to be allocated to the colour ramp height}

\item{zlim}{minimum and maximum values for the colour ramp}

\item{cm}{object of class 'colourmap' to plot (or default spatstat colour ramp if left null)}

\item{steps}{how many numerical steps to label in the ramp legend}

\item{sigdigits}{how to round the digits in the labels}

\item{rounding}{to how many decimal places should the labels be rounded}

\item{vertical}{whether the ramp is places as a vertical or horizontal ribbon}
}
\value{
an image ribbon added to an existing plot.
}
\description{
Handle the positioning and labelling of a (spatstat) colour ramp more easily.
}
\examples{
exampledensitymap <- density(cells, 0.05)
plot(exampledensitymap, ribbon=FALSE)
ribbonplot(x=0.1, y=0.1, xprop=0.03, yprop=0.2, zlim=c(min(exampledensitymap),max(exampledensitymap)), cex.axis=0.6, col.axis="white", col.ticks="white", las=2)
}
