% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{scaleBar}
\alias{scaleBar}
\title{Produce a very simple map scale-bar}
\usage{
scaleBar(x, y, width, height, ...)
}
\arguments{
\item{x}{Top left x coordinate}

\item{y}{Top left y coordinate}
}
\value{
a plotted polygon
}
\description{
This function draw a simple scale-bar in the form of a box with a certain width and height (in map units) and hung on its top right corner.
}
