% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcomponents.R
\name{spEllipse}
\alias{spEllipse}
\title{Produce an circle or ellipse as a SpatialPolygons object}
\usage{
spEllipse(xc, yc, a, b = a, angle = 0, nsteps = 360,
  proj4string = CRS(as.character(NA)), noSp = FALSE, ...)
}
\arguments{
\item{xc}{x coordinate of circle or ellipse centre}

\item{yc}{y coordinate of circle or ellipse centre}

\item{a}{the size of the circle radius in map units or the length of the longer axis of the ellipse}

\item{b}{the size of the shorter axis of the ellipse, if necessary (default is circular with b=a).}

\item{angle}{the map orientation of the shorter axis of the ellipse, if necessary (default is longer axis is oriented north).}

\item{nsteps}{the number of vertices used to describe the circle of the ellipse, if necessary (i.e. the smoothness of the plotted circle, default is 360)}
}
\value{
a plotted polygon
}
\description{
This function draw a simple circle or ellipse, centred on a certain coordinate pair and of a certain size (the circle's radius or height, width and azimuth for an ellipse). The smooothness of the circle can be controlled by the chosen number of vertices placed along the perimeter, and the circle can be assigned a projection system if desired.
}
