% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonprocessing.R
\name{voronoi}
\alias{voronoi}
\title{Simple and weighted Voronoi tesselations}
\usage{
voronoi(x, ids, weights = NULL, win = NULL, maxdist = NULL,
  nsteps = 36, verbose = TRUE, ...)
}
\arguments{
\item{x}{an object of class SpatialPoints*}

\item{ids}{identifiers for each point as lables for the resulting polygons}

\item{weights}{optional weight values for the calculation (e.g. measure of relative `influence' or `size' of each point)}

\item{win}{an object of class SpatialPolygons* for the exterior of the study area.}

\item{maxdist}{A cut-off (the radius of a circle) for the extent of a point's tesselation.}

\item{nsteps}{How many increments to use for defining the cruvature of a circle.}

\item{verbose}{A logical variable indicating whether extra information on progress should be reported. Default is TRUE.}
}
\value{
An object of class SpatialPolygonsDataFrame
}
\description{
Function to produce a Voronoi tesselation (aka Dirichelet set or Thiessen polygons) from a set of point, with or without weights.
}
\examples{
w <- SpatialPolygons(list( Polygons(list(Polygon(cbind(c(0,0,100,100,0), c(0,100,100,0,0)))), "1")), 1:1)
pts <- data.frame(X=runif(3)*100,Y=runif(3)*100,ptID=letters[1:3], Size=c(1,2,4))
coordinates(pts) <- ~X+Y
vor <- voronoi(pts, ids=pts$ptID, win=w)
vorw <- voronoi(pts, ids=pts$ptID, weights=pts$Size, win=w)
plot(vor, lty="dotted", border="red")
plot(vorw, add=TRUE)
points(pts)
}
