% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertformats.R
\name{xyTransform}
\alias{xyTransform}
\title{Convert a coordinate pair to a different coordinate system}
\usage{
xyTransform(coords, proj1, proj2)
}
\arguments{
\item{coords}{A vector of length 2 with the coordinates}

\item{proj1}{The CRS of the input coordinates}

\item{proj2}{The target CRS of the output coordinates}
}
\value{
A vector of length 2 with the converted coordinate pair.
}
\description{
Converts a single coordinate pair to a different coordinate system
}
\examples{
bng <- CRS("+init=epsg:27700") # British National Grid
lonlat <- CRS("+init=epsg:4326") # LatLon WGS84
xyTransform(c(345600,378200), bng, lonlat)
}
